Allow smart mirror to gracefully continue operation without fitbit integration if no fitbit configuration in config.js.
This commit allows smart mirror to start up and not display fitbit data if there is no fitbit config defined. The service won't be called in this case and no errors or debug console messages appear either, so this allows for graceful operation. If the config is added for fitbit, then fitbit stats are displayed on the smart mirror UI.

Configuration for fitbit is all handled in config.js now if a user would like to use it.

PR details for this change are located: https://github.com/evancohen/smart-mirror/pull/180