Remove fs from package json
You may not be aware of the recent chaos around the 'fs' module, but believe me you don't want it in here. In short, all it does is print `I am fs`. You probably installed it looking for `require('fs')` which is a native module and not an NPM one.

See http://status.npmjs.org/incidents/dw8cr1lwxkcr